… .claude/commands/workflow/task-dev.md
… backend/app/core/base_request_utils.py
… backend/app/core/base_security.py
… backend/app/core/base_validators.py
… backend/app/main.py
… backend/app/middleware/audit_action_detector.py
… backend/app/middleware/audit_middleware.py
… backend/app/models/audit.py
… backend/app/models/audit_schemas.py
… backend/app/routers/audit.py
… backend/app/security/audit_security.py
… backend/app/security/rate_limiter.py
… backend/app/services/audit.py
… backend/app/services/audit_queries.py
… backend/app/services/audit_service_utils.py
… backend/create_test_db.py
… backend/migrations/env.py
… backend/migrations/versions/002_create_audit_logs.py
… backend/migrations/versions/003_optimize_audit_indexes.py
… backend/test_5_second_audit.py
… backend/tests/conftest.py
… backend/tests/fixtures/audit_fixtures.py
… backend/tests/test_audit_integration.py
… backend/tests/test_audit_performance.py
… backend/tests/test_audit_security.py
… backend/tests/test_config_api.py
… backend/validate_security.py
… cypress/e2e/audit-logs.cy.ts
… cypress/fixtures/audit-test-data.json
… cypress/support/audit-commands.ts